pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bricklayer"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9",
    "pyyaml>=6.0",
]

[project.scripts]
bricklayer = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*"]
